follow_up_bot/requirements.txt
flask
gunicorn
anthropic
python-dotenv
requests
apscheduler
pytz
psycopg2-binary
